Extract authentication code to separate helper
This is now generic enough to use anywhere.
